code/backend/libraries/myDependencies/requirements/requirements.txt
aiomysql==0.2.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
boto3==1.38.17
botocore==1.38.17
certifi==2025.4.26
click==8.2.0
fastapi==0.115.12
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
jmespath==1.0.1
myAws @ file:///Users/juanasoto/Desktop/JUMP/blossom/code/backend/libraries/myAws
myEncryption @ file:///Users/juanasoto/Desktop/JUMP/blossom/code/backend/libraries/myEncryption
myExceptions @ file:///Users/juanasoto/Desktop/JUMP/blossom/code/backend/libraries/myExceptions
myOrm @ file:///Users/juanasoto/Desktop/JUMP/blossom/code/backend/libraries/myOrm
packaging==25.0
pluggy==1.6.0
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.10.1
PyMySQL==1.1.1
pytest==8.3.5
pytest-asyncio==0.26.0
python-dateutil==2.9.0.post0
s3transfer==0.12.0
setuptools==80.7.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
wheel==0.45.1
